Remove initstate from index to books Readucer
